chore: refactor model api validation to visitor func pattern
