+++ schema/testdata/errors/api_duplicate_definitions.keel
@@ -3,8 +3,9 @@ model Test {
 
 api Web {
     models {
+        //expect-error:9:13:DuplicateDefinitionError:The model 'Test' has already been included in the 'Web' API
         Test
-        //expect-error:9:13:E021:Model 'Test' has already been included in the Web API
+        //expect-error:9:13:DuplicateDefinitionError:The model 'Test' has already been included in the 'Web' API
         Test
     }
 }
